+++ tests/test_hetnode2vec_tf.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 import os.path
 from xn2v import CSFGraph
-from xn2v.hetnode2vec_tf import N2vGraph
+from xn2v.hetnode2vec import N2vGraph
 from tests.utils.utils import calculate_total_probs
